mingw32 deprecated, adapt to never version of MSYS2

+++ lib/select.c
@@ -531,11 +531,11 @@ restart:
         {
           /* Perform handle->descriptor mapping.  */
           WSAEventSelect ((SOCKET) h, NULL, 0);
-          if (FD_ISSET (h, &handle_rfds))
+          if (FD_ISSET ((unsigned long long)h, &handle_rfds))
             FD_SET (i, rfds);
-          if (FD_ISSET (h, &handle_wfds))
+          if (FD_ISSET ((unsigned long long)h, &handle_wfds))
             FD_SET (i, wfds);
-          if (FD_ISSET (h, &handle_xfds))
+          if (FD_ISSET ((unsigned long long)h, &handle_xfds))
             FD_SET (i, xfds);
         }
       else
